+++ lib/td_auth/permissions/permission.ex
@@ -51,7 +51,8 @@ defmodule TdAuth.Permissions.Permission do
     manage_ingest_relations: "manage_ingest_relations",
     view_data_structures_profile: "view_data_structures_profile",
     view_quality_rule: "view_quality_rule",
-    manage_quality_rule_implementations: "manage_quality_rule_implementations"
+    manage_quality_rule_implementations: "manage_quality_rule_implementations",
+    execute_quality_rule: "execute_quality_rule"
   }
 
   schema "permissions" do