+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased] 
+
+### Changed
+
+- [TD-2002] Update td-cache and delete permissions list from config
+
 ## [3.1.0] 2019-07-08
 
 ### Changed