+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## [Unreleased]
+## [Unreleased] 
 
+### Changed
+
+- [TD-2002] Update td-cache and delete permissions list from config
 - [TD-1776] add permission view_quality_rule
 
 ## [3.1.0] 2019-07-08